feat: moved simulator
