+++ apps/web/src/components/workflow/Preview/common/phone-simulator/AppleIcon.tsx
