+++ apps/web/src/components/workflow/Preview/index.ts
@@ -1,3 +1,4 @@
 export * from './chat';
 export * from './Email';
 export * from './push';
+export * from './sms';
