+++ apps/web/src/components/workflow/Preview/sms/index.ts
@@ -1,0 +1,1 @@
+export * from './SmsPreview';
